# api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Isidor API",
    "description": "\n    # Isidor Health Optimization API\n    \n    Isidor is an AI-driven life protocol system designed to optimize fitness, nutrition, sleep, cognitive performance, and overall well-being.\n    \n    ## Key Features\n    \n    * **Health Data Integration**: Track and analyze health metrics from various sources\n    * **Protocol System**: Follow structured frameworks for health optimization\n    * **AI-Powered Insights**: Get personalized health insights using Gemini AI\n    * **Vector Embeddings**: Utilize semantic search for finding patterns in health data\n    * **Privacy-First Design**: Secure storage of health data with encryption\n    \n    ## Authentication\n    \n    This API uses JWT-based authentication. To access protected endpoints:\n    \n    1. Obtain an access token using the `/api/v1/auth/login` endpoint\n    2. Include the token in the Authorization header: `Authorization: Bearer {token}`\n    3. Use the refresh token to obtain a new access token when it expires\n    \n    ## Rate Limiting\n    \n    API requests are rate-limited to prevent abuse. The current limit is 100 requests per hour per IP address.\n    \n    ## Data Privacy\n    \n    All sensitive health data is encrypted at rest using PostgreSQL's pgcrypto extension.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ai/protocols/{user_id}": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Get Protocol Recommendation",
        "description": "Generate a protocol recommendation using Gemini AI.",
        "operationId": "get_protocol_recommendation_api_v1_ai_protocols__user_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolRecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Protocol Recommendation Api V1 Ai Protocols  User Id  Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/protocols/": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Read Protocols",
        "description": "Retrieve protocols.",
        "operationId": "read_protocols_api_v1_protocols__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "target_metric",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Protocol"
                  },
                  "title": "Response Read Protocols Api V1 Protocols  Get"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Protocols"
        ],
        "summary": "Create Protocol Endpoint",
        "description": "Create new protocol.",
        "operationId": "create_protocol_endpoint_api_v1_protocols__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/protocols/{protocol_id}": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Read Protocol",
        "description": "Get protocol by ID.",
        "operationId": "read_protocol_api_v1_protocols__protocol_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Protocols"
        ],
        "summary": "Update Protocol Endpoint",
        "description": "Update a protocol.",
        "operationId": "update_protocol_endpoint_api_v1_protocols__protocol_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Protocols"
        ],
        "summary": "Delete Protocol Endpoint",
        "description": "Delete a protocol.",
        "operationId": "delete_protocol_endpoint_api_v1_protocols__protocol_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Protocol Endpoint Api V1 Protocols  Protocol Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/protocols/templates/list": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Read Protocol Templates",
        "description": "Retrieve available protocol templates.",
        "operationId": "read_protocol_templates_api_v1_protocols_templates_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProtocolTemplate"
                  },
                  "type": "array",
                  "title": "Response Read Protocol Templates Api V1 Protocols Templates List Get"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/protocols/templates/create": {
      "post": {
        "tags": [
          "Protocols"
        ],
        "summary": "Create From Template",
        "description": "Create a protocol from a template.",
        "operationId": "create_from_template_api_v1_protocols_templates_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolTemplateCustomization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/protocols/templates/{template_id}/details": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Get Template Details",
        "description": "Get detailed information about a protocol template.",
        "operationId": "get_template_details_api_v1_protocols_templates__template_id__details_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Template Details Api V1 Protocols Templates  Template Id  Details Get"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Protocol": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "target_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Metrics"
          },
          "duration_type": {
            "type": "string",
            "title": "Duration Type"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_metrics",
          "duration_type",
          "id"
        ],
        "title": "Protocol",
        "description": "Protocol schema for response."
      },
      "ProtocolCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "target_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Metrics"
          },
          "duration_type": {
            "type": "string",
            "title": "Duration Type"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_metrics",
          "duration_type"
        ],
        "title": "ProtocolCreate",
        "description": "Protocol creation schema."
      },
      "ProtocolRecommendationRequest": {
        "properties": {
          "health_goal": {
            "type": "string",
            "title": "Health Goal"
          },
          "current_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Current Metrics"
          }
        },
        "type": "object",
        "required": [
          "health_goal",
          "current_metrics"
        ],
        "title": "ProtocolRecommendationRequest"
      },
      "ProtocolTemplate": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "target_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Metrics"
          },
          "duration_type": {
            "type": "string",
            "title": "Duration Type"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "name",
          "description",
          "target_metrics",
          "duration_type"
        ],
        "title": "ProtocolTemplate",
        "description": "Protocol template schema."
      },
      "ProtocolTemplateCustomization": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "target_metrics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Metrics"
          },
          "duration_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Type"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "ProtocolTemplateCustomization",
        "description": "Protocol template customization schema."
      },
      "ProtocolUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "target_metrics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Metrics"
          },
          "duration_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Type"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          }
        },
        "type": "object",
        "title": "ProtocolUpdate",
        "description": "Protocol update schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}